Continue expanding the server install
This really needs to be split to a separate deb package, but in the
interest of using this as a template formula, I am continuing on with
the hardest path.

I have had to give up on the scgi setup as getting the CPAN SCGI
installed is just going far to overboard.  The salt cpan is a module
only, and it has bugs if run on bare system.   ie.. root/.cpan/cache
doesn't exist if you have never run cpan, and the module borks on that.
Obviously the person who made the module already had that directory :)

Now to decide how to hand off the Apache, or NGINX, or Whatever config
(or do we even bother?)