Totally redo the testing to make it more fine grained - rsync only
I was having fresh out of the box test run failures do to a race
condition (not yet resolved) that was leaving artificacts, so the second
run would fail.   Decided to ramp up the testing module to do more fine
grained salt calls with the state id individually.   Side effect is we
get to do a lot more tests that handling various state branches for
requirements, onchanges, etc.